chore: move header to AppHeader file
